fixcode

+++ BusinessObject/DTO/AgencyPromotionRequestDto.cs
@@ -11,9 +11,9 @@ namespace BusinessObject.DTO
         public Guid AgencyPromotionRequestId { get; set; }
         public long AgencyId { get; set; }
         public string AgencyName { get; set; }
-        public int CurrentLevelId { get; set; }
-        public int SuggestedLevelId { get; set; }
-        public int TotalScore { get; set; }
+        public long CurrentLevelId { get; set; }
+        public long SuggestedLevelId { get; set; }
+        public long TotalScore { get; set; }
         public string Status { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? ReviewedAt { get; set; }
