+++ BusinessObject/DTO/AgencyScoreDetailDto.cs
@@ -10,7 +10,7 @@ namespace BusinessObject.DTO
     {
         public long AgencyId { get; set; }
         public string AgencyName { get; set; }
-        public int TotalScore { get; set; }
+        public long TotalScore { get; set; }
         public List<AgencyScoreItemDto> ScoreHistory { get; set; }
     }
 }
